Move code to separate module to avoid circular import
